re-run to get results consistent with private repo
This is the bane of my existence, but for reasons that I still don't understand the exact numerical values fit by the global epistasis model seem to slightly fluctuate depending on which core on the rhino computing cluster the jobs run on. The differences are always very small and I have no idea why this happens (different numerical precision on different cores?).

In any case, all of the manually fine-tuned figures that Tyler and I have made used the values obtained running the prior analysis in the private repo for all Abs, and then the first run here had them slightly differently. So re-ran a few times until the cluster hit a node that gave values matching those from before even out to the fifth decimal place.